requirements.txt
grpcio==1.26.0
grpcio-tools==1.26.0
numpy==1.18.1
protobuf==3.11.2
opencv-python==4.1.0.25
Flask==1.1.1